Move implementation for ShapeInfo to templates.h file
